dvc-ad-intel/requirements.txt
langchain
playwright
python-dotenv
google-cloud-videointelligence
psycopg2-binary
pg8000
